imglib2-style: change line length to 140 characters
This change is made in order to better support the long class names that
are part of imglib2 like "RandomAccessibleInterval" .

The long class names easily produces long lines of code. Breaking these
long lines with a code formatter leads to less readable code. The line
width is increased here in order to avoid a fight between developers and
code autoformatting :)