imglib2-style: enable @formatter:on and @formatter:off tags
This commit configures the formatter on/off tags.
Previously the setting seemed to have been ignored. With the new version
of the formatter settings XML we need to set the values properly, to make
the formatter:on/off tags work.

A comment like this in the java source code, turns off the formatting off
for the following lines: // @formatter:off
Formatting can be turned on again with: // @formatter:on